fixe 2 (i forgot you can't mix int and int64)

+++ muxhandlers/game.go
@@ -500,7 +500,7 @@ func QuickPostGameResults(helper *helper.Helper) {
 		}
 
 		// increase character(s)'s experience
-		expIncrease := 0
+		expIncrease := int64(0)
 		if gameconf.CFile.UseRingsAsExp {
 			expIncrease = request.Rings + request.FailureRings // all rings collected
 		} else {
@@ -871,12 +871,12 @@ func PostGameResults(helper *helper.Helper) {
 		}
 
 		// increase character(s)'s experience
-		expIncrease := 0
+		expIncrease := int64(0)
 		if gameconf.CFile.UseRingsAsExp {
 			expIncrease = request.Rings + request.FailureRings // all rings collected
 		} else {
 			expIncrease = request.Score / 5000
-			if request.BossDestroyed != 0 {
+			if request.BossDestroyed != int64(0) {
 				// TODO: check to see if this is right! we might need to do additional checks here to make sure the game actually wants to move onto the next stage, since if it's on the same mileage point assume the boss isn't defeated and set XP to 0
 				expIncrease = 2000
 			}
